+++ RedLineКUnitTest/RedLineUnitTest.csproj
@@ -49,8 +49,9 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="DetailTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ComputationsTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
