--- RedLineКUnitTest/UnitTest1.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using RedLineProject;
-
-namespace RedLineКUnitTest
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-            Assert.AreEqual(0, 0);
-        }
-    }
-}
